p4i integration Integrate 261518 from 4.3 to main: Fixes:    use double precision in ECMA parseFloat() implementation RevBy:    TrustMe AutoTest: 50 more tests now pass in the SpiderMonkey test suite Details:  should be pretty self explanatory; a rather silly bug, just because the ECMA function has the word "Float" in it does not mean we should use QString::toFloat() :-) (in ECMA, all floating point numbers are double precision)
[git-p4: depot-paths = "//depot/qt/main/": change = 261519]

+++ src/script/qscriptecmaglobal.cpp
@@ -406,7 +406,7 @@ QScriptValueImpl Global::method_parseFloat(QScriptContextPrivate *context,
 
     QString str = context->argument(0).toString().trimmed();
     bool ok = false;
-    qsreal result = str.toFloat(&ok);
+    qsreal result = str.toDouble(&ok);
     if (!ok) {
         if (str == QLatin1String("Infinity"))
             result = +qInf();
